requirements.txt
opencv-python==4.8.1.78
deepface==0.0.79
openai==0.28.0
python-dotenv==1.0.0
numpy==1.24.3
sounddevice==0.4.6
soundfile==0.12.1
matplotlib==3.7.2
flask==2.3.3
webrtcvad==2.0.10
scipy==1.11.2
pandas==2.1.0